// src/scene/scene-intro/actors/robocillo/robocillo-animations.d.ts
export declare enum RobocilloAnimations {
    STOP_SIDE = 0,
    PAPER_TAKE = 1,
    PAPER_CHECK = 2,
    PAPER_THROW = 3,
    SIDE_TO_FRONT = 4,
    FRONT_TO_SIDE = 5,
    STOP_FRONT = 6,
    WALK = 7,
    MOVE_HANDS = 8,
    RAISE_HANDS = 9,
    JUMP_FRONT = 10
}
export declare enum RobocilloKeyFrames {
    FLOOR_CONTACT = 0
}